requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-Mail==0.9.1
python-dotenv==0.19.0
PyJWT==2.3.0
Werkzeug==2.0.1
pymysql==1.0.2
gunicorn==20.1.0          # Needed for Render to serve Flask
apscheduler==3.9.1        # For BackgroundScheduler
